Test: adicionando testes da camada de infraestrutura

+++ simulacao/app/src/main/java/dev/enzosoares/consignado/simulacao/simulacaoemprestimo/dataprovider/SimulacaoEmprestimoMapper.java
@@ -4,21 +4,19 @@ import dev.enzosoares.consignado.simulacao.cliente.valueobject.CPF;
 import dev.enzosoares.consignado.simulacao.cliente.valueobject.Convenio;
 import dev.enzosoares.consignado.simulacao.simulacaoemprestimo.SimulacaoEmprestimo;
 
-import java.math.BigDecimal;
-
 public class SimulacaoEmprestimoMapper {
 
     public static SimulacaoEmprestimoJpaEntity toEntity(SimulacaoEmprestimo simulacaoEmprestimo) {
         return new SimulacaoEmprestimoJpaEntity(
-        simulacaoEmprestimo.getIdSimulacao().toString(),
-        simulacaoEmprestimo.getDataSimulacao(),
-        simulacaoEmprestimo.getCpfCliente().getValue(),
-        simulacaoEmprestimo.getConvenioCliente().getKey(),
-        simulacaoEmprestimo.getValorSolicitado().doubleValue(),
-        simulacaoEmprestimo.getTaxaJuros().doubleValue(),
-        simulacaoEmprestimo.getQuantidadeParcelas(),
-        simulacaoEmprestimo.getValorParcela().doubleValue(),
-        simulacaoEmprestimo.getValorEmprestimo().doubleValue()
+                simulacaoEmprestimo.getIdSimulacao().toString(),
+                simulacaoEmprestimo.getDataSimulacao(),
+                simulacaoEmprestimo.getCpfCliente().getValue(),
+                simulacaoEmprestimo.getConvenioCliente().getKey(),
+                simulacaoEmprestimo.getValorSolicitado(),
+                simulacaoEmprestimo.getTaxaJuros(),
+                simulacaoEmprestimo.getQuantidadeParcelas(),
+                simulacaoEmprestimo.getValorParcela(),
+                simulacaoEmprestimo.getValorEmprestimo()
         );
     }
 
@@ -28,11 +26,11 @@ public class SimulacaoEmprestimoMapper {
                 simulacaoEmprestimoEntity.getDataSimulacao(),
                 CPF.with(simulacaoEmprestimoEntity.getCpf()),
                 Convenio.getConvenioByKey(simulacaoEmprestimoEntity.getConvenio()),
-                BigDecimal.valueOf(simulacaoEmprestimoEntity.getValorParcela()),
-                BigDecimal.valueOf(simulacaoEmprestimoEntity.getValorTotal()),
-                BigDecimal.valueOf(simulacaoEmprestimoEntity.getTaxaAplicada()),
+                simulacaoEmprestimoEntity.getValorParcela(),
+                simulacaoEmprestimoEntity.getValorTotal(),
+                simulacaoEmprestimoEntity.getTaxaAplicada(),
                 simulacaoEmprestimoEntity.getQuantidadeParcelas(),
-                BigDecimal.valueOf(simulacaoEmprestimoEntity.getValorSolicitado())
-                );
+                simulacaoEmprestimoEntity.getValorSolicitado()
+        );
     }
 }
